90 percent done

+++ src/controllers/user.controller.ts
@@ -3,6 +3,7 @@ import { userLoginValidator, userRegisterValidator, userUpdateValidator } from '
 import prisma from "../db"
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
+
 export async function UserLogin(req: Request, res: Response) {
     try {
         const body = req.body;
@@ -10,26 +11,23 @@ export async function UserLogin(req: Request, res: Response) {
         if (!check.success) {
             res.status(400).json({
                 success: false,
-                message: check.error
+                message: check.error.message
             });
             return
         }
-        const user = await prisma.user.findFirst({
-            where: {
-                email: check.data.email,
-            }
-        });
+
+        const user = await prisma.user.findFirst({ where: { email: check.data.email } });
         if (!user) {
-            res.json({
+            res.status(404).json({
                 success: false,
                 message: "User not found"
             });
             return
         }
-        const checkPassword = await bcrypt.compare(check.data.password, user.password);
 
+        const checkPassword = await bcrypt.compare(check.data.password, user.password);
         if (!checkPassword) {
-            res.json({
+            res.status(401).json({
                 success: false,
                 message: "Password is incorrect"
             });
@@ -37,20 +35,19 @@ export async function UserLogin(req: Request, res: Response) {
         }
 
         const token = jwt.sign({ id: user.id }, process.env.JWT_SECRET!);
-        res.json({
+        res.status(200).json({
             success: true,
             message: "User logged in successfully",
             data: token
         });
         return
-    } catch (error) {
-        res.json({
+    } catch (error: any) {
+        res.status(500).json({
             success: false,
-            message: error
+            message: error.message
         });
         return
     }
-
 }
 
 export async function UserRegister(req: Request, res: Response) {
@@ -58,34 +55,26 @@ export async function UserRegister(req: Request, res: Response) {
         const body = req.body;
         const check = userRegisterValidator.safeParse(body);
         if (!check.success) {
-            res.json({
+            res.status(400).json({
                 success: false,
-                message: check.error
+                message: check.error.message
             });
             return
         }
-        const checkUser = await prisma.user.findFirst({
-            where: {
-                email: check.data.email,
-            }
-        });
-
+        const checkUser = await prisma.user.findFirst({ where: { email: check.data.email } });
         if (checkUser) {
-            res.json({
+            res.status(409).json({
                 success: false,
                 message: "User already exists"
             });
             return
         }
-        const checkDomain = await prisma.user.findFirst({
-            where: {
-                domain: check.data.domain
-            }
-        });
+
+        const checkDomain = await prisma.user.findFirst({ where: { domain: check.data.domain } });
         if (checkDomain) {
-            res.json({
+            res.status(409).json({
                 success: false,
-                message: "doamin is taken"
+                message: "Domain is taken"
             });
             return
         }
@@ -98,18 +87,19 @@ export async function UserRegister(req: Request, res: Response) {
                 password: hashedPassword,
                 domain: check.data.domain
             }
-        })
+        });
+
         const token = jwt.sign({ id: user.id }, process.env.JWT_SECRET!);
-        res.json({
+        res.status(201).json({
             success: true,
             message: "User registered successfully",
             data: token
         });
         return
-    } catch (error) {
-        res.status(400).json({
+    } catch (error: any) {
+        res.status(500).json({
             success: false,
-            message: error
+            message: error.message
         });
         return
     }
@@ -119,16 +109,15 @@ export async function GetProfile(req: Request, res: Response) {
     try {
         const domain = req.query.domain;
         if (!domain) {
-            res.json({
+            res.status(400).json({
                 success: false,
                 message: "Domain is required"
             });
             return
         }
+
         const profile = await prisma.user.findUnique({
-            where: {
-                domain: domain.toString()
-            },
+            where: { domain: domain.toString() },
             include: {
                 experience: true,
                 skills: true,
@@ -136,53 +125,57 @@ export async function GetProfile(req: Request, res: Response) {
                 projects: true,
                 style: true,
             }
-        })
-        res.json({
+        });
+
+        if (!profile) {
+            res.json({
+                success: false,
+                message: "Profile not found"
+            });
+            return
+        }
+
+        res.status(200).json({
             success: true,
             message: "Profile fetched successfully",
             data: profile
         });
         return
     } catch (error: any) {
-        res.json({
+        res.status(500).json({
             success: false,
             message: error.message
         });
         return
-
     }
 }
 
 export async function UpdateProfile(req: Request, res: Response) {
     try {
-        const userId = req.userId!
-        const body = req.body
-        console.log(userId)
+        const userId = req.userId!;
+        const body = req.body;
+
         const check = userUpdateValidator.safeParse(body);
         if (!check.success) {
-            res.json({
+            res.status(400).json({
                 success: false,
-                message: check.error
+                message: check.error.message
             });
             return
         }
+
         await prisma.user.update({
-            where: {
-                id: userId
-            },
-            data: {
-                bio: check.data.bio,
-                domain: check.data.domain
-            }
-        })
-        res.json({
+            where: { id: userId },
+            data: { bio: check.data.bio, domain: check.data.domain }
+        });
+
+        res.status(200).json({
             success: true,
             message: "Profile updated successfully"
         });
         return
     } catch (error: any) {
-        console.log(error)
-        res.json({
+        res.status(500).json({
             success: false,
             message: error.message
         });
@@ -192,19 +185,16 @@ export async function UpdateProfile(req: Request, res: Response) {
 
 export async function DeleteProfile(req: Request, res: Response) {
     try {
-        const userId = req.userId
-        await prisma.user.delete({
-            where: {
-                id: userId
-            }
-        })
-        res.json({
+        const userId = req.userId;
+        await prisma.user.delete({ where: { id: userId } });
+
+        res.status(200).json({
             success: true,
             message: "Account deleted"
-        })
+        });
         return
     } catch (error: any) {
-        res.json({
+        res.status(500).json({
             success: false,
             message: error.message
         });
